pyproject.toml
[project]
name = "ai-agents-crewai-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai = {extras = ["tools"], version = "^0.85.0"}
python-dotenv = "^1.0.1"
langchain-community = "^0.3.19"



[tool.poetry]
packages = [{include = "ai_agents_crewai_template", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
run-agents = "ai_agents_crewai_template.main:run"